feat(web): add get messages route
